Sort stocktake update into respective folder/files
